Bug Fix PopUpWindow Locking Player and helper
Bug Fix: Player was moving when interacting with the inventory, which shouldnt happen. It happend that the openPopUpWIndow script set the inventory.InteractionWithUIActive variable to lock the player to false everytime in its update method if the popupwindow was not open. This was not right so an event was implemented for the popupwindow in the resumePanel to set inventory.InteractionWithUIActive to false only when the popup window was closed.
Helper: Animation was added and adjusted.